requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
fpdf2>=2.7.9
pyhanko>=0.21.0
python-multipart>=0.0.9
